+++ appsTransfer/src/main/java/org/kinecosystem/appstransfer/presenter/SenderServiceBinder.kt
@@ -25,6 +25,7 @@ class SenderServiceBinder(private val context: Context?) {
         fun onServiceDisconnected()
         fun onTransferFailed(errorMessage: String, senderAddress: String)
         fun onTransferComplete(kinTransferComplete: SendKinServiceBase.KinTransferComplete)
+        fun onTransferTimeout()
     }
 
     private val mainThreadHandler = Handler(Looper.getMainLooper())
@@ -34,6 +35,11 @@ class SenderServiceBinder(private val context: Context?) {
         private set
     private var transferService: SendKinServiceBase? = null
     private var listener: BinderListener? = null
+    private val handler = Handler()
+    private var afterTimeout = false
+    @Volatile
+    private var transferResponseReceived = false
+
     private val connection = object : ServiceConnection {
 
         override fun onServiceConnected(className: ComponentName, service: IBinder) {
@@ -73,11 +79,12 @@ class SenderServiceBinder(private val context: Context?) {
         if (isBounded) {
             executorService.execute {
                 try {
+                    startTimeoutCounter()
                     val kinTransferComplete: SendKinServiceBase.KinTransferComplete? = transferService?.transferKin(receiverAddress, amount, memo)
                     kinTransferComplete?.let {
                         updateTransferCompleted(it)
                     } ?: kotlin.run {
-                        startSendKinAsnyc(receiverAddress, amount, memo, object : KinTransferCallback {
+                        startSendKinAsync(receiverAddress, amount, memo, object : KinTransferCallback {
                             override fun onSuccess(kinTransferComplete: SendKinServiceBase.KinTransferComplete) {
                                 updateTransferCompleted(kinTransferComplete)
                             }
@@ -97,7 +104,7 @@ class SenderServiceBinder(private val context: Context?) {
         }
     }
 
-    fun startSendKinAsnyc(receiverAddress: String, amount: Int, memo: String, callback: KinTransferCallback) {
+    private fun startSendKinAsync(receiverAddress: String, amount: Int, memo: String, callback: KinTransferCallback) {
         if (isBounded) {
             executorService.execute {
                 transferService?.transferKinAsync(receiverAddress, amount, memo, callback)
@@ -147,16 +154,37 @@ class SenderServiceBinder(private val context: Context?) {
         }
     }
 
+    private fun startTimeoutCounter() {
+        afterTimeout = false
+        transferResponseReceived = false
+        handler.postDelayed({
+            updateTransferTimeout()
+        }, Consts.TRANSACTION_TIMEOUT)
+    }
+
+    private fun updateTransferTimeout() {
+        afterTimeout = true
+        if (!transferResponseReceived) {
+            listener?.onTransferTimeout()
+        }
+    }
+
     private fun updateTransferCompleted(kinTransferComplete: SendKinServiceBase.KinTransferComplete) {
-        mainThreadHandler.post {
-            listener?.onTransferComplete(kinTransferComplete)
+        transferResponseReceived = true
+        if (!afterTimeout) {
+            mainThreadHandler.post {
+                listener?.onTransferComplete(kinTransferComplete)
+            }
         }
     }
 
     private fun updateTransferFailed(e: SendKinServiceBase.KinTransferException) {
-        mainThreadHandler.post {
-            val msg = e.message ?: ""
-            listener?.onTransferFailed(msg, e.senderAddress)
+        transferResponseReceived = true
+        if (!afterTimeout) {
+            mainThreadHandler.post {
+                val msg = e.message ?: ""
+                listener?.onTransferFailed(msg, e.senderAddress)
+            }
         }
     }
